chore!: upgrade to cdktf 0.20.0 (#298)
This PR initiates the upgrade of CDKTF from version `0.19.0` to version
`0.20.0`.
Unfortunately, not everything can be automated, and the following steps
need to be completed manually:
 
- [x] Update `@cdktf/provider-null` to a version compatible with
`cdktf@0.20.0`
[here](https://github.com/cdktf/cdktf-tf-module-stack/blob/106036d2d6b65652b68dabb2648b1f7980c6421d/.projenrc.ts#L78).
Look up the version
[here](https://github.com/cdktf/cdktf-provider-null/releases/).
- [x] Update `@cdktf/provider-random` to a version compatible with
`cdktf@0.20.0`
[here](https://github.com/cdktf/cdktf-tf-module-stack/blob/106036d2d6b65652b68dabb2648b1f7980c6421d/.projenrc.ts#L79).
Look up the version
[here](https://github.com/cdktf/cdktf-provider-random/releases/).
- [x] Run `npx projen`
 
Please checkout this PR, complete the above steps, push the changes to
this branch, and then mark this PR as ready for review to complete the
upgrade. Thanks!

---------

Signed-off-by: team-tf-cdk <github-team-tf-cdk@hashicorp.com>
Co-authored-by: Ansgar Mertens <ansgar@hashicorp.com>